+++ asyncgit/src/graph/chunk.rs
@@ -6,8 +6,8 @@ pub enum Markers {
 
 #[derive(Clone, Debug)]
 pub struct Chunk {
-	pub alias: Option<u32>,
-	pub parent_a: Option<u32>,
-	pub parent_b: Option<u32>,
+	pub alias: Option<usize>,
+	pub parent_a: Option<usize>,
+	pub parent_b: Option<usize>,
 	pub marker: Markers,
 }
